+++ tests/callbacks/test_device_stats_monitor.py
@@ -17,6 +17,7 @@ import pytest
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import DeviceStatsMonitor
+from pytorch_lightning.callbacks.device_stats_monitor import _prefix_metric_keys
 from pytorch_lightning.loggers import CSVLogger
 from pytorch_lightning.utilities.distributed import rank_zero_only
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
@@ -128,3 +129,12 @@ def test_device_stats_monitor_no_logger(tmpdir):
 
     with pytest.raises(MisconfigurationException, match="Trainer that has no logger."):
         trainer.fit(model)
+
+
+def test_prefix_metric_keys(tmpdir):
+    """Test that metric key names are converted correctly."""
+    metrics = {"1": 1.0, "2": 2.0, "3": 3.0}
+    prefix = "foo"
+    separator = "."
+    converted_metrics = _prefix_metric_keys(metrics, prefix, separator)
+    assert converted_metrics == {"foo.1": 1.0, "foo.2": 2.0, "foo.3": 3.0}